% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_realtime.R
\name{gtfs_realtime}
\alias{gtfs_realtime}
\title{GTFS real time}
\usage{
gtfs_realtime(response)
}
\arguments{
\item{response}{binary response body}
}
\description{
Parses the response from a GTFS real-time feed
}

