% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_realtime.R
\name{gtfs_tripUpdates}
\alias{gtfs_tripUpdates}
\title{GTFS Trip Updates}
\usage{
gtfs_tripUpdates(FeedMessage)
}
\arguments{
\item{FeedMessage}{returned from \link{gtfs_realtime}}
}
\value{
list of \code{data.table} trip update tables
}
\description{
Returns a list of trip update information given a FeedMessage input
}
\examples{
\dontrun{

## South-East Queensland
url <- "https://gtfsrt.api.translink.com.au/Feed/SEQ"

response <- httr::GET(url,
httr::accept_json(),
httr::add_headers('Authorization' = ''))

lst <- gtfs_tripUpdates(gtfs_realtime(response, content = "FeedMessage"))
}

}

