% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{get_gtfs}
\alias{get_gtfs}
\title{Get gtfs}
\usage{
get_gtfs(url)
}
\arguments{
\item{url}{link to GTFS .zip}
}
\value{
list of dataframes, one for each .txt file
}
\description{
Downloads and extracts GTFS static data
}

