% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_realtime.R
\name{gtfs_realtime}
\alias{gtfs_realtime}
\title{GTFS real time}
\usage{
gtfs_realtime(response, content = c("FeedMessage", "Alert", "EntitySelector",
  "FeedEntity", "FeedHeader", "VehiclePosition", "TimeRange",
  "TranslateedString", "TripDescriptor", "TripUpdate",
  "TripUpdate.StopTimeEvent", "TripUpdate.StopTimeUpdate", "VehicleDescriptor",
  "Vehicle"))
}
\arguments{
\item{response}{binary response body}

\item{content}{String specifying which section of the response to return}
}
\value{
FeedMessage RProtoBuf object containing the feed message
}
\description{
Parses the raw response from a GTFS real-time feed
}
\examples{
\dontrun{
## South-East Queensland
url <- "https://gtfsrt.api.translink.com.au/Feed/SEQ"

response <- httr::GET(url,
httr::accept_json(),
httr::add_headers('Authorization' = ''))

feed <- gtfs_realtime(response, content = "FeedMessage")

}
}

